external services fixtures
